Changed the button text to icons
Removed the button text and added icons
Set the onhover text to the former button text
Added CSS to style the button

Only didn't get the import button with the right CSS because this is not a real button, but just a div. Looks like the padding might be the problem there.